requirements.txt
confluent_kafka==2.3.0
icecream==2.1.3
python-dotenv==1.0.0